# Core: Keep track of data files to be removed for orphaned DV detection