@nastra 2025-07-08
thanks for the reviews @amogh-jahagirdar and @stevenzwu. I'll go ahead and merge this to unblock the release and we can follow up on https://github.com/apache/iceberg/pull/13222#discussion_r2191655872 if necessary